+++ contentcuration/contentcuration/frontend/shared/Base.vue
