+++ docs/architecture.md
@@ -17,10 +17,11 @@
 
 Studio is currently divided into multiple single-page-applications - each one has a distinct function and purpose. It is plausible that one day, this will be collapsed into one SPA to rule them all, but today is not that day.
 
-The main SPAs currently are:
+Examples of SPAs in Studio are:
 
 * Channel List - responsible for showing over all listings of channels under different categories (public, editable, etc.) and also for creating new channels, and creating and editing collections of channels.
 * Channel Edit - responsible for showing the contents of an individual channel and allowing a user with the appropriate permissions to edit the contents and structure of that channel, it is the primary workhorse of the Studio user interface.
+* Accounts - responsible for creating accounts, logging in, "forgot password" functionality, and other related items
 * Administration - an admin only page that allows people with staff privileges to see extensive information about user accounts, user created channels, and other detailed information in order to support users in using Studio.
 * Settings - allows a user to modify their user account settings, this could potentially be included in the other SPAs rather than having its own SPA.
 
@@ -35,9 +36,16 @@ All code for SPAs should live in `frontend/<spaName>`. For example, the Channel 
 
 Code that is common across multiple SPAs is in the [`frontend/shared`](../contentcuration/contentcuration/frontend/shared) folder. This can be referenced with the [Webpack](https://webpack.js.org/concepts/) alias `shared/...` in ES6 imports.
 
-Currently, some code is still housed inside [`static/js`](../contentcuration/contentcuration/static/js). This should be migrated into the `frontend` folder as soon as possible. As a temporary convenience during this migration, anything in the frontend folder can also be referenced with `frontend/...`. Relative imports are to be preferred within the SPA folders.
+(Note: currently, some code is still housed inside [`static/js`](../contentcuration/contentcuration/static/js). This should be migrated into the `frontend` folder as soon as possible. As a temporary convenience during this migration, anything in the frontend folder can also be referenced with `frontend/...`. Relative imports are to be preferred within the SPA folders.)
 
-Within each of the `frontend` folders, Vuex modules are stored in `vuex`, Vue components are stored in `views` and the Vuex store, [Vue Router](https://router.vuejs.org/guide/), and entry point for the application are all defined in the root of the relevant SPA folder.
+By convention, within each of the `frontend` sub-folders:
+
+* `index.js` contains code to initialize the Vue application
+* `router.js` stores client-side [Vue Router](https://router.vuejs.org/guide/) routes
+* The `pages` directory contains components associated with routes
+* The `components` directory contains components shared across multiple pages in the SPA. (Components that are only used within a single page should go in a sub-directory for that page.)
+* If there are any custom layouts for the SPA, they would go in a `layouts` directory
+* Vuex modules are stored in the `vuex` folder and the store is initialized in `store.js`
 
 Tests should be defined in line with Jest conventions inside a folder named `__tests__` inside the same folder as the code they are testing. This allows the use of relative imports to quickly access the relevant code. Some efforts have been made to systematically setup the test environment to handle most scenarios in the *root* folder [`/jest_config/setup.js`](../jest_config/setup.js), so any test should not need to setup any Vue plugins such as Vuex or Vue Router.
